Add VariantDependencies support to ModelBuilder v2
Created a new GlobalLibraryBuildService. As soon as Studio
moves to v2, the old one can go away.

Expanded on the test fixtures to dump VariantDependencies
and GlobalLibraryMap and compare to golden files outside
of the test class (reaching kotlin compiler limit on size
of raw strings!)

Also update the model action and the return types to
handle either AndroidProject or VariantDependencies

Bug: N/A
Test: Expanded BasicModelV2Test and added AppModelTest
Change-Id: Icd9599361f415846019f515d8ed64b48c6c9ca9c