160626532: LintClient.isGradle() called before clientName initialization
This CL switches the clientName over to a lateinit property.
This ensures that if any code attempts to access the clientName
before the client has been configured, it will throw an error.

This was particularly important because in some cases we were
accessing the client name during issue configuration (to for example
enable/disable issues if we're running in Studio or Gradle), and
issue registries are often configured *before* creating the LintClient,
which is where until now the client name was configured.

Now, the various entry points for lint (such as the AGP setup, and
the IDE setup), initializes the client name property as the first
step.

(This CL also restores the RequiredAttributeDetector when running
in the IDE. This was primarily broken from AGP, when using
checkDependencies=false. There is now a Gradle integration test
which verifies that this check is indeed off when run from Gradle.)

Fixes: 160626532
Test: Existing + new unit test included

Change-Id: Ia641c5240185d68376421cfdc4b4cd392ae38527