Package uncompressed dex files by default for minSdk >= 28
If dex is compressed in an apk, it will get uncompressed and stored in
vdex.  If dex is uncompressed in an apk, ART (since P) knows it can find
it in the apk, and won't store an extra copy in vdex.

Previously, we would only package uncompressed dex files if
useEmbeddedDex was set to true in the merged manifest. With this change,
we package uncompressed dex files if useEmbeddedDex is true OR if minSdk
>= 28.

Bug: 161461387
Test: PackagingUtilsTest, UseEmbeddedDexPackagingTest
Change-Id: Ifa3944fdc13421b3274c045550c26726ca4964a2