Lint: fix perf regression (checkDependencies=true)
After the UastEnvironment rewrite, we now delegate to
KotlinCoreEnvironment to create PackagePartProviders. Unfortunately,
KotlinCoreEnvironment.createPackagePartProvider() permanently appends
the PackagePartProvider to the environment, which significantly slows
down resolve over time as the environment is reused.

The root issue is that KotlinCoreEnvironment was not designed to be
used repeatedly for multiple analyses---which we currently do when
checkDependencies=true. Fortunately, there is work underway to move
to a model where the UastEnvironment is used to analyze only one
module, at which point we can remove the workaround added here.

Lint performance on the Tivi project using a warm Gradle daemon:
    Before UastEnvironment rewrite: 1m 23s
    After UastEnvironment rewrite: 3m 47s
    After this change: 1m 14s

Bug: 153758215
Test: manual (none of our build benchmarks use checkDependencies=true)
Change-Id: Ib1e93d09fcb43fcc9227b6ee38897db7e6c6a906