Stop collecting coverage from classes without location data
Including classes without a source location in Jacoco output no longer
seems to be necessary. Comparing the reports shows a negligible
difference of only 9 net covered lines.

The motivation for this change is the migration from JDK8 to JDK11.
As is referenced in https://github.com/jacoco/jacoco/issues/16#issuecomment-351943050
inclnolocationclasses=true causes issues starting with JDK9 and requires
an explicit exclclassloader argument for 9+. Using the default
inclnolocationclasses=false avoids the issue and is safe both pre- and
post- migration.

As including no-location classes provides negligible improvement,
I'm just going to go with the simpler option for the migration.

Test: N/A
Fixes: 161616412
Change-Id: I66f1191e3c7d99b0d01fe99cc4410d4b0b8671ca