Add app-metadata.properties to APKs and .aab files
This change creates a new AppMetadataTask which creates an
app-metadata.properties file which is included in APKs and app bundle
files.

The app-metadata.properties file currently just contains the app
metadata format version (1.0) and the AGP version. The motivation for
including this file in APKs and app bundles is to be able to track which
AGP version created these files in the play store.

Bug: 148861019
Test: AppMetadataTaskTest unit and integration tests
Change-Id: Ib5eb0d9cd7fa2266cc647ce127bea2bbae5abda0