Remove storeType from builder model.
This is not used for signing, but only for creating the
debug keystore when it's missing. The customization of this
value is very unlikely, and really does not belong in the "signing"
config.

This may become configurable later via property or DSL (unlikely
since it's very rare) but this probably does not need to be in the
PSD (which would be the only reason to have it in the model.

This does allow us to remove the fork between the JDK8 and JDK11
dumps.

Bug: 161919627
Test: updated model dumps
Change-Id: Iff93ac4085d9620f2fd1f7a9917d971973cad808