Attempts to improve/refactor GradleTestProject.
The class is huge and tons a ton around creating the test
project, interacting with it, discovering sdks, repos,
etc...

This is a first step that moves some static values
that cannot be static.
Information about build location, gradleUserHome
may change depending on the test when running under Bazel
and therefore must be computed based on each instance instead.

Moves all that static init to new classes to make the main
class less gigantic.

There is still some mess around SDK present in multiple version
as `androidHome` and `androidSdkHome` that needs to be fixed
in a separate CL.

There's a lot more changes to clean things up, but one
has to go slowly to keep things easy to understand and
reviewable.

Next up is also removing the content generation into its
own class and make the API from inside the test more
clear as right now there are lots of items that are confusing
(e.g. getSubProject is not related to the sub-project created
via TestProject, and yet they receive lots of properties
mostly needed by the creation of the files).

Bug: N/A
Test: existing tests
Change-Id: I079a766a7841e919a09041f9b5062d1012dd9412