[lint,Dsl] rework GroovyGradleVisitor
Instead of walking the statements in a block by hand, use the fact
that we will visit all method calls because we visit TupleExpressions,
and that the existence and hierarchy of blocks is stored in
the method call stack we maintain.

When checking a method call, if it is of the form that might be
considered setting a Dsl property in the lint-check sense (a single
unnamed argument or a single map of named arguments), check it for
property aspects as well as method call aspects.

Bug: 161877341
Test: existing (no change in functionality)
Change-Id: Ib82fb55f2126daf93397e5248db86a981bbf028f