Lint: workaround for Kotlin 1.4 update
Lint unit tests run using the same Kotlin stdlib version that is used
in Studio. However, Lint gets the Kotlin language level from
//prebuilts/tools/common/lint-psi, which is updated separately.

We are currently in a situation where Studio is moving to Kotlin 1.4,
but lint-psi is still at 1.3.70. This breaks LocaleDetectorTest
because there is a method call which resolves to the wrong stdlib
declaration due to the fact that the correct declaration is annotated
with @SinceKotlin("1.4").

This will not be a problem once lint-psi is updated to Kotlin 1.4.
In the meantime, we work around the issue by explicitly setting
the Kotlin langauge level to at least 1.4.

Bug: 162855232
Test: fixes LocaleDetectorTest
Change-Id: I855ed9987444a69bc15f62d38f41f3a6572c1de6